pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shimexpy"
version = "0.1.0"
description = "An open-source Python package for spatial harmonics X-ray imaging."
readme = "readme.md"
authors = [{name = "Jorge Luis Beltran Diaz"}]
license = { text = "Apache-2.0" }
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
]

dependencies = [
    "numpy",
    "xarray",
    "dask",
    "matplotlib",
    "scikit-image",
    "tifffile",
    "scipy"
]

[project.optional-dependencies]
dev = ["pytest", "setuptools", "build", "twine"]
gui = ["PySide6>=6.6"]
cuda13x = ["cupy-cuda13x"]
cuda12x = ["cupy-cuda12x"]
cuda11x = ["cupy-cuda11x"]

[project.scripts]
shimexpy = "shimexpy.gui.app:run_gui"
shi = "shimexpy.cli.cli:run_cli"

[tool.setuptools.packages.find]
where = ["src"]
include = ["shimexpy*"]

[tool.setuptools.exclude-package-data]
"*" = ["__pycache__/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "E501", "W503"]
